Synching between users and tasks collection

+++ src/routers/task.js
@@ -2,10 +2,13 @@ const express=require('express')
 const router=new express.Router()
 const Task=require('../models/task')
 const brcyprt=require('bcryptjs')
+const auth=require('../middleware/auth.js')
 
-router.post('/tasks',async(req,res)=>{
-    const task=new Task(req.body)
-
+router.post('/tasks',auth,async(req,res)=>{
+    const task=new Task({
+        ...req.body,
+        owner:req.user._id
+    })
     try{
         await task.save()
         res.status(201).send(task)
@@ -15,22 +18,23 @@ router.post('/tasks',async(req,res)=>{
 })
 
 
-router.get('/tasks',async(req,res)=>{
+router.get('/tasks',auth,async(req,res)=>{
 
     try{
-        const tasks=await Task.find({})
-        res.send(tasks)
+        // const tasks=await Task.find({owner:req.user._id})
+        await req.user.populate('tasks').execPopulate()
+        res.send(req.user.tasks)
     }catch(e){
         res.status(500).send()
     }
 
 })
 
-router.get('/tasks/:id',async(req,res)=>{
+router.get('/tasks/:id',auth,async(req,res)=>{
     const _id=req.params.id
 
     try{
-        const task=await Task.findById(_id)
+        const task=await Task.findOne({_id,owner:req.user._id})
         if(!task){
             return res.status(400).send()
         }
@@ -41,8 +45,7 @@ router.get('/tasks/:id',async(req,res)=>{
 })
 
 
-router.patch('/tasks/:id',async(req,res)=>{
-    const _id=req.params.id
+router.patch('/tasks/:id',auth,async(req,res)=>{
 
     const updates=Object.keys(req.body)
     const allowedUpdates=['status','description']
@@ -57,20 +60,17 @@ router.patch('/tasks/:id',async(req,res)=>{
 
     try{
 
-        const task=await Task.findById(_id)
+        const task=await Task.findById({_id:req.params.id,owner:req.user._id})
 
         updates.forEach((update)=>{
             task[update]=req.body[update]
         })
 
-        await task.save()
-
-        // const task= await Task.findByIdAndUpdate(_id,req.body,{new:true,runValidators:true})
-
         if(!task){
-           return res.status(404).send()
-        }
+            return res.status(404).send()
+         }
 
+        await task.save()
         res.send(task)
     }catch(e){
         res.status(400).send(e)
@@ -78,11 +78,11 @@ router.patch('/tasks/:id',async(req,res)=>{
 })
 
 
-router.delete('/tasks/:id',async(req,res)=>{
+router.delete('/tasks/:id',auth,async(req,res)=>{
     const _id=req.params.id
 
     try{
-        const task=await Task.findByIdAndDelete(_id)
+        const task=await Task.findByIdAndDelete({_id:req.params.id,owner:req.user._id})
 
         if(!task){
             return res.status(404).send()
